+++ apps/backend/src/app/api/latest/integrations/neon/internal/confirm/route.tsx
@@ -18,7 +18,7 @@ export const POST = createSmartRouteHandler({
     body: yupObject({
       interaction_uid: yupString().defined(),
       project_id: yupString().defined(),
-      neon_project_name: yupString().optional(),
+      external_project_name: yupString().optional(),
     }).defined(),
   }),
   response: yupObject({
@@ -33,7 +33,7 @@ export const POST = createSmartRouteHandler({
     const set = await globalPrismaClient.apiKeySet.create({
       data: {
         projectId: req.body.project_id,
-        description: `Auto-generated for Neon${req.body.neon_project_name ? ` (${req.body.neon_project_name})` : ""}`,
+        description: `Auto-generated for ${req.body.external_project_name ? `"${req.body.external_project_name}"` : "an external project"}`,
         expiresAt: new Date(Date.now() + 1000 * 60 * 60 * 24 * 365 * 100),
         superSecretAdminKey: `sak_${generateSecureRandomString()}`,
       },