+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @stackframe/e2e-tests
 
+## 2.8.15
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/js@2.8.15
+  - @stackframe/stack-shared@2.8.15
+
+## 2.8.14
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.14
+  - @stackframe/js@2.8.14
+
+## 2.8.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.13
+  - @stackframe/js@2.8.13
+
 ## 2.8.12
 
 ### Patch Changes