+++ examples/react-example/CHANGELOG.md
@@ -1,5 +1,24 @@
 # react-example
 
+## 2.8.15
+
+### Patch Changes
+
+- @stackframe/react@2.8.15
+
+## 2.8.14
+
+### Patch Changes
+
+- @stackframe/react@2.8.14
+
+## 2.8.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/react@2.8.13
+
 ## 2.8.12
 
 ### Patch Changes