+++ packages/init-stack/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @stackframe/init-stack
 
+## 2.8.15
+
+### Patch Changes
+
+- @stackframe/stack-shared@2.8.15
+
+## 2.8.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.14
+
+## 2.8.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.13
+
 ## 2.8.12
 
 ### Patch Changes