+++ packages/stack-sc/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @stackframe/stack-sc
 
+## 2.8.15
+
+## 2.8.14
+
+## 2.8.13
+
+### Patch Changes
+
+- Various changes
+
 ## 2.8.12
 
 ## 2.8.11